backend/requirements.txt
asgiref==3.8.1
certifi==2026.2.25
charset-normalizer==3.4.6
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.11
pillow==12.1.1
psycopg2-binary==2.9.11
PyJWT==2.8.0
python-dotenv==1.0.1
razorpay==2.0.1
requests==2.32.5
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.6.3
